로그인 시 password의 검증 로직 추가 작성, id의 데이터 타입을 Long으로 변경

+++ src/main/java/io/github/authservice/crowdfund/feature/auth/response/SignUpResponse.java
@@ -2,6 +2,6 @@ package io.github.authservice.crowdfund.feature.auth.response;
 
 public record SignUpResponse(
         String message,
-        Integer userId
+        Long userId
 ) {
 }
