+++ src/main/java/io/github/authservice/crowdfund/feature/useraddress/response/UpdateUserAddressResponse.java
@@ -2,6 +2,6 @@ package io.github.authservice.crowdfund.feature.useraddress.response;
 
 public record UpdateUserAddressResponse(
         String message,
-        Integer addressId
+        Long addressId
 ) {
 }
